pyproject.toml
[tool.poetry]
name = "atalakit-infrastructure"
version = "0.0.0"
description = "IaC for Atalakit"
readme = "README.md"
authors = ["Stephen Palfreyman"]

[tool.poetry.dependencies]
python = "^3.11"
boto3 = "*"
pyyaml = "*"
python-terraform = "*"

[tool.poetry.group.dev.dependencies]
bandit = "*"
black = "*"
pytest = "*"
pytest-black = "*"
pytest-cov = "*"
reorder-python-imports = "*"

[build-system]
requires = ["poetry_core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
